ci: new github-action for git-l10n code review
Repository of git-l10n is a fork from "git/git" on GitHub, and uses
GitHub pull request for code review. A helper program "git-po-helper"
can be used to check typos in ".po" files, validate syntax, and check
commit messages. It would be convenient to integrate this helper program
to CI and add comments in pull request.

The new github-action workflow is added in ".github/workflows/". It will
only be enabled for the following cases:

 * A repository named as "git-po", such as a repository forked from
   "git-l10n/git-po".

 * Push to the branch that contains "l10n" in the name.

 * Pull reqeust from a remote branch which has "l10n" in the name, such
   as: "l10n/fix-fuzzy-translations".

The new l10n workflow listens to two types of github events:

    on: [push, pull_request_target]

The reason we use "pull_request_target" instead of "pull_request" is
that for security reasons, pull requests from forks receive a read-only
GITHUB_TOKEN and workflows cannot write comments back to pull requests.
GitHub provides a "pull_request_target" event to resolve security risks
by checking out the base commit from the target repository, and provide
write permissions for the workflow.

By default, adminstrators can set strict permissions for workflows. The
following code is used to modify the permissions for the GITHUB_TOKEN
and give write permission in order to create comments in pull-requests.

    permissions:
      pull-requests: write

This workflow will scan commits one by one. If a commit does not look
like a l10n commit (no file in "po/" has been changed), the scan process
will stop immediately. For a "push" event, no error will be reported
because it is normal to push non-l10n commits merged from upstream. But
for the "pull_request_target" event, errors will be reported.
For this reason, additional option is provided for "git-po-helper".

    git-po-helper check-commits \
        --github-action-event="${{ github.event_name }}" -- \
        <base>..<head>

The output messages of "git-po-helper" contain color codes not only for
console, but also for logfile. This is because "git-po-helper" uses a
package named "logrus" for logging, and I use an additional option
"ForceColor" to initialize "logrus" to print messages in a user-friendly
format in logfile output. These color codes help produce beautiful
output for logs of workflow, but they must be stripped off when creating
comments for pull requests. E.g.:

    perl -pe 's/\e\[[0-9;]*m//g' git-po-helper.out

"git-po-helper" may generate two kinds of suggestions, errors and
warnings. All the errors and warnings will be reported in the log
the l10n workflow. For a "pull_request_target" event, an additional
comment will be created in the pull request to report the result.
A l10n contributor should try to fix all the errors, and should pay
attention to the warnings.

Helped-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>